# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Royal Midnight luxury black car service API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List all drivers
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [drivers]
      summary: Register a driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDriverBody"
      responses:
        "201":
          description: Driver created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get driver by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDriver
      tags: [drivers]
      summary: Update driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverBody"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}/toggle-availability:
    patch:
      operationId: toggleDriverAvailability
      tags: [drivers]
      summary: Toggle driver online/offline status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleAvailabilityBody"
      responses:
        "200":
          description: Driver updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}/earnings:
    get:
      operationId: getDriverEarnings
      tags: [drivers]
      summary: Get driver earnings summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Earnings summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverEarnings"

  # ── USERS ──────────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ── PROMO CODES ────────────────────────────────────────────────────────────
  /promos:
    get:
      operationId: listPromos
      tags: [promos]
      summary: List promo codes
      responses:
        "200":
          description: Promo codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromoCode"
    post:
      operationId: createPromo
      tags: [promos]
      summary: Create a promo code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromoBody"
      responses:
        "201":
          description: Promo created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCode"
  /promos/{id}:
    patch:
      operationId: updatePromo
      tags: [promos]
      summary: Update promo code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePromoBody"
      responses:
        "200":
          description: Updated promo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCode"
    delete:
      operationId: deletePromo
      tags: [promos]
      summary: Delete promo code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /promos/validate:
    post:
      operationId: validatePromo
      tags: [promos]
      summary: Validate and apply a promo code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidatePromoBody"
      responses:
        "200":
          description: Promo validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoValidationResult"

  # ── PRICING RULES ──────────────────────────────────────────────────────────
  /pricing:
    get:
      operationId: listPricingRules
      tags: [pricing]
      summary: List pricing rules
      responses:
        "200":
          description: Pricing rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PricingRule"
    post:
      operationId: createPricingRule
      tags: [pricing]
      summary: Create a pricing rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePricingRuleBody"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingRule"
  /pricing/{id}:
    patch:
      operationId: updatePricingRule
      tags: [pricing]
      summary: Update pricing rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePricingRuleBody"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingRule"
    delete:
      operationId: deletePricingRule
      tags: [pricing]
      summary: Delete pricing rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ADMIN ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Driver:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        licenseNumber:
          type: string
        status:
          type: string
          enum: [pending, active, inactive, suspended]
        isOnline:
          type: boolean
        rating:
          type: ["number", "null"]
        totalRides:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, phone, licenseNumber, status, isOnline, totalRides, createdAt]
    CreateDriverBody:
      type: object
      properties:
        userId:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        licenseNumber:
          type: string
      required: [name, email, phone, licenseNumber]
    UpdateDriverBody:
      type: object
      properties:
        status:
          type: ["string", "null"]
          enum: [pending, active, inactive, suspended, null]
        rating:
          type: ["number", "null"]
    ToggleAvailabilityBody:
      type: object
      properties:
        isOnline:
          type: boolean
      required: [isOnline]
    DriverEarnings:
      type: object
      properties:
        totalEarnings:
          type: number
        thisMonth:
          type: number
        thisWeek:
          type: number
        today:
          type: number
        totalRides:
          type: integer
        avgPerRide:
          type: number
        recentPayouts:
          type: array
          items:
            $ref: "#/components/schemas/EarningsPeriod"
      required: [totalEarnings, thisMonth, thisWeek, today, totalRides, avgPerRide, recentPayouts]
    EarningsPeriod:
      type: object
      properties:
        date:
          type: string
        amount:
          type: number
        rides:
          type: integer
      required: [date, amount, rides]

    # ── BOOKING SCHEMAS ───────────────────────────────────────────────────────
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [booking, driver, payment, system, promo]
        isRead:
          type: boolean
        bookingId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, title, message, type, isRead, createdAt]

    # ── PROMO SCHEMAS ─────────────────────────────────────────────────────────
    PromoCode:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minBookingAmount:
          type: ["number", "null"]
        maxUses:
          type: ["integer", "null"]
        usedCount:
          type: integer
        isActive:
          type: boolean
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, code, description, discountType, discountValue, usedCount, isActive, createdAt]
    CreatePromoBody:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minBookingAmount:
          type: ["number", "null"]
        maxUses:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
      required: [code, description, discountType, discountValue]
    UpdatePromoBody:
      type: object
      properties:
        isActive:
          type: ["boolean", "null"]
        expiresAt:
          type: ["string", "null"]
    ValidatePromoBody:
      type: object
      properties:
        code:
          type: string
        bookingAmount:
          type: number
      required: [code, bookingAmount]
    PromoValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        discountAmount:
          type: ["number", "null"]
        finalAmount:
          type: ["number", "null"]
        message:
          type: string
      required: [valid, message]

    # ── PRICING RULE SCHEMAS ──────────────────────────────────────────────────
    PricingRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        vehicleClass:
          type: ["string", "null"]
        baseFare:
          type: number
        ratePerMile:
          type: number
        airportSurcharge:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, baseFare, ratePerMile, airportSurcharge, isActive, createdAt]
    CreatePricingRuleBody:
      type: object
      properties:
        name:
          type: string
        vehicleClass:
          type: ["string", "null"]
        baseFare:
          type: number
        ratePerMile:
          type: number
        airportSurcharge:
          type: number
      required: [name, baseFare, ratePerMile, airportSurcharge]
    UpdatePricingRuleBody:
      type: object
      properties:
        baseFare:
          type: ["number", "null"]
        ratePerMile:
          type: ["number", "null"]
        airportSurcharge:
          type: ["number", "null"]
        isActive:
          type: ["boolean", "null"]

    # ── ADMIN SCHEMAS ─────────────────────────────────────────────────────────
...