# lib/api-zod/src/generated/types/toggleAvailabilityBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Royal Midnight luxury black car service API
 * OpenAPI spec version: 0.2.0
 */

export interface ToggleAvailabilityBody {
  isOnline: boolean;
}